+++ CHANGES.md
@@ -2,9 +2,13 @@
 
 ## [Unreleased] - ReleaseDate
 
+### Changes
+
+- Added `child_index` method to get the index of a child in a parent.
+
 ## [0.19.0] - 2024-01-15
 
-### Change
+### Changes
 
 - Add `Debug` support for `Xot` struct.
 
@@ -454,8 +458,8 @@
 Initial public release.
 
 <!-- next-url -->
-[Unreleased]: https://github.com/faassen/xot/compare/v0.19.0...HEAD
 
+[Unreleased]: https://github.com/faassen/xot/compare/v0.19.0...HEAD
 [0.19.0]: https://github.com/faassen/xot/compare/v0.18.0...v0.19.0
 [0.18.0]: https://github.com/faassen/xot/compare/v0.17.0...v0.18.0
 [0.17.0]: https://github.com/faassen/xot/compare/v0.16.0...v0.17.0